Fixed
added coins collection.
////////////
fixed the projectile firing.
