+++ Assets/Lucas/Coins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ea59752092a57d4b93d4fd67733bfa6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
