+++ Assets/Lucas/Coins/Prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 32b29af869c88384cb00d399b841a2e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
