+++ Assets/Lucas/Coins/Prefab/Coin.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d39078df11868b94cbb58575aa65687a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
