+++ Assets/Lucas/Coins/Prefab/Coins Canvas.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b3802cee38d6f5b409cb3c7204c1a67d
-DefaultImporter:
+guid: 490d339dc2ff07d4eb3ce3f043d62e82
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
