+++ Assets/Lucas/Coins/Scripts.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: b9d789f2f24ab8b4cb086b9716dba746
+guid: a127e72bcce09a34483c9b465a6134d5
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
