+++ Assets/Lucas/Coins/Scripts/CoinsColect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c134a00035ac924587956fb2c52b2c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
