+++ Assets/Lucas/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90f6efc03b723c74fba3f5a054891718
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
