+++ Assets/Lucas/Player/Sprites/projectile Basic.png.meta
