--- Assets/Lucas/Scripts/Projectile.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Projectile : MonoBehaviour
-{
-    //Speed
-    public float Speed = 30;
-
-    //rigidBody
-    private Rigidbody2D RB;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        RB = GetComponent<Rigidbody2D>();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        RB.AddForce(Vector2.up * Speed, ForceMode2D.Impulse);
-    }
-
-}
